Revert "Add mongo as event"
