+++ QueriesApp/Services/MongoService.cs
@@ -1,9 +1,16 @@
 using MongoDB.Driver;
 
-public class MongoAsDocumentService : MongoServiceBase
+public class MongoService : ServiceBase<MongoPOCO>
 {
-    public MongoAsDocumentService(int quantity) : base(quantity)
+    private IMongoCollection<MongoPOCO> db { get; set; }
+    public MongoService(int quantity) : base(quantity)
     {
+        db = new MongoContext().POCO;
+    }
+
+    protected override void DoSelect(SelectQuery s)
+    {
+        var _ = db.Find(_ => _.AccountId == s.AccountId).FirstOrDefault();
     }
 
     protected override void ExecuteInserts(IEnumerable<MongoPOCO> pocos)
@@ -50,4 +57,15 @@ public class MongoAsDocumentService : MongoServiceBase
         if (newDocuments.Any())
             db.InsertMany(newDocuments);
     }
+
+    protected override MongoPOCO MakePoco()
+        => new MongoPOCO
+        {
+            AccountId = Random.Next(AccountsRange[0], AccountsRange[1]),
+            Statement = [ new MongoStatement
+            {
+                Value = double.Round(Random.NextDouble() * ValuesRange[Random.Next(0, 2)], 2),
+                Timestamp = Random.NextInt64(TimestampRange[0], TimestampRange[1])
+            }]
+        };
 }
